HSEARCH-1549 Moving the helper tests under util.impl to match the location of the classes they test
